pyproject.toml
[tool.poetry]
name = "veloapi-stuff"
version = "0.1.0"
description = ""
authors = ["Nick Barrett <nick.barrett@broadcom.com>"]
readme = "README.md"
packages = [
    { include = "veloapi" }
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.6"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
websockets = "^13.1"
pyyaml = "^6.0.2"
aiostream = "^0.6.2"
dataclasses-json = "^0.6.7"
jsondiff = "^2.2.1"
ijson = "^3.3.0"
duckdb = "^1.2.0"
polars = "^1.22.0"
pyarrow = "^19.0.0"
fastapi = {extras = ["standard"], version = "^0.115.8"}
apscheduler = "^3.11.0"
dataclass-csv = "^1.4.0"
geopy = "^2.4.1"
jsonpatch = "^1.33"
intervaltree = "^3.1.0"
matplotlib = "^3.10.1"
pandas = "^2.2.3"
pysnmp = "^7.1.16"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"